Refactor and unify internal exports
