+++ docs/data/toolpad/core/components/dashboard-layout/DashboardLayoutFullScreen.tsx
@@ -3,7 +3,7 @@ import { createTheme } from '@mui/material/styles';
 import MapIcon from '@mui/icons-material/Map';
 import { AppProvider, type Navigation } from '@toolpad/core/AppProvider';
 import { DashboardLayout } from '@toolpad/core/DashboardLayout';
-import { useDemoRouter } from '@toolpad/core/internals';
+import { useDemoRouter } from '@toolpad/core/internal';
 
 const NAVIGATION: Navigation = [
   {
