+++ docs/data/toolpad/core/components/page-container/BasicPageContainer.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { PageContainer } from '@toolpad/core/PageContainer';
 import { AppProvider } from '@toolpad/core/AppProvider';
-import { useDemoRouter } from '@toolpad/core/internals';
+import { useDemoRouter } from '@toolpad/core/internal';
 import { useTheme } from '@mui/material/styles';
 import Paper from '@mui/material/Paper';
