+++ packages/toolpad-core/src/internal/context.ts
@@ -1,0 +1,3 @@
+import * as React from 'react';
+
+export const DocsContext = React.createContext<boolean>(false);
