+++ packages/toolpad-core/src/internal/demo.tsx
@@ -30,5 +30,3 @@ export function useDemoRouter(initialUrl: string = '/') {
 
   return router;
 }
-
-export { Link } from '../shared/Link';
