--- packages/toolpad-core/src/internal/index.ts
@@ -1,3 +1,0 @@
-import { DocsContext as Docs } from '../shared/context';
-
-export const DocsContext = Docs;
