+++ packages/toolpad-core/src/internal/index.tsx
@@ -1,0 +1,4 @@
+export { Link } from '../shared/Link';
+
+export * from './demo';
+export { DocsContext } from './context';
