--- packages/toolpad-core/src/internals/index.tsx
@@ -1,1 +1,0 @@
-export * from './demo';
