+++ packages/toolpad-core/src/shared/context.ts
@@ -19,5 +19,3 @@ export const PaletteModeContext = React.createContext<{
 export const RouterContext = React.createContext<Router | null>(null);
 
 export const WindowContext = React.createContext<Window | undefined>(undefined);
-
-export const DocsContext = React.createContext<boolean>(false);
